convert all money columns to cents - Changed 18 currency columns across 8 tables from RealColumn (dollars) to IntColumn (cents) - Updated all related providers, display formatting, and input parsing to handle cents - Added versioned backup support (export_format_version 2) with proper dollars-to-cents conversion on v1 imports - Fixed seed data defaults - Minor cleanup (unused import, etc.) 18june26

+++ lib/presentation/providers/timer_providers.dart
@@ -62,7 +62,7 @@ class TimerActions extends Notifier<AsyncValue<void>> {
     required int employeeId,
     int? costCodeId,
     String? workDetails,
-    double? hourlyRate,
+    int? hourlyRate,
     DateTime? startTime,
   }) async {
     state = const AsyncValue.loading();
@@ -88,7 +88,7 @@ class TimerActions extends Notifier<AsyncValue<void>> {
     required int employeeId,
     int? costCodeId,
     String? workDetails,
-    double? hourlyRate,
+    int? hourlyRate,
     required DateTime startTime,
   }) async {
     state = const AsyncValue.loading();
